Revert dependencies to Yarn and update
The initial commit included a yarn.lock file, so it seems like this
project has been using Yarn from the beginning. The instructions in the
readme have mentioned npm instead of Yarn for a long time, and the
previous commit added a package-lock.json; this commit updates the
readme to recommend Yarn, removes the package-lock.json, and updates the
dependencies in yarn.lock.